+++ src/main/java/org/opentripplanner/standalone/config/NodeAdapter.java
@@ -100,6 +100,10 @@ public class NodeAdapter {
         return result;
     }
 
+    public boolean isNonEmptyArray() {
+        return json.isArray() && json.size() > 0;
+    }
+
     public String getSource() {
         return source;
     }